pyproject.toml
[project]
name = "pykanban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ctkmenubar>=0.8",
    "customtkinter>=5.2.2",
    "hpyt>=1.3.4",
    "pillow>=11.0.0",
    "pywinstyles>=1.8",
]
